geneweb-python/requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
pydantic
python-multipart
httpx
pytest-mock
itsdangerous
pycodestyle
pip-audit